feat: share rate limiter across auth routes
